Rename staticCompleter to inkCompleter
And move it to its own file.
